#207 paginate cli interface
